Update README.md
Some additions and amendments that are largely cosmetic. The main functional edit is about the end-game scoring - I propose that the individual with the highest score wins, so that it is still possible for the Draw player to win even if there is a checkmate, by virtue of having taken substantially more pieces than anyone else. Otherwise they will be heavily disadvantaged against teams with many people.